[BE] 관심지역 등록

See merge request s08-bigdata-recom-sub2/S08P22D106!59

+++ server/aro/api-module/src/main/java/com/nassafy/api/service/InterestService.java
@@ -39,8 +39,10 @@ public class InterestService {
 
         // 해당 회원 관심 지역 비우기
         List<Interest> interests = member.getInterests();
-        for (Interest interest : interests) {
-            interestRepository.delete(interest);
+        if(interests != null) {
+            for (Interest interest : interests) {
+                interestRepository.delete(interest);
+            }
         }
 
         // 영속성 컨텍스트에 반영
